Resolved conflict in reprompt():
- Use main's version with $rp_pending counter instead of early return
- This tracks pending reprompts while busy, allowing them to be
  executed when no longer busy
- Removed duplicate $busy declaration

+++ Smak.pm
@@ -131,7 +131,9 @@ our $remote_cd = '';  # Remote directory for SSH workers
 our $job_server_socket;  # Socket to job-master
 our $job_server_pid;  # PID of job-master process
 our $job_server_master_port;  # Master port for reconnection
-our $busy = 0;  # Set when a build is in progress
+
+# Output control
+our $stomp_prompt = "\r";
 
 sub set_report_mode {
     my ($enabled, $fh) = @_;
@@ -2544,6 +2546,7 @@ sub unified_cli {
     my $quiet = $opts{opts} || 1;
 
     my $prompt = $opts{prompt} || 'smak> ';
+    my $ret = "detach";
 
     # Track state
     my $watch_enabled = 0;
@@ -2731,35 +2734,39 @@ sub unified_cli {
     $interactive = 0;
 
     # Cleanup on exit
-    if ($exit_requested) {
+    if ($detached) {
+        # Detach was requested (explicit detach command only, not Ctrl-C)
+        $SmakCli::cli_owner = -1;  # Mark CLI as unowned
+        if ($own_server) {
+            print "Detaching from CLI (job server $Smak::job_server_pid still running)...\n";
+        } else {
+            print "Detaching from job server...\n";
+        }
+    }
+    elsif ($exit_requested) {
         # Always shut down the server when quitting (even when attached)
         if ($socket) {
             print "Shutting down job server...\n";
             print $socket "SHUTDOWN\n";
             my $ack = <$socket>;
         }
-    } elsif ($detached) {
-        # Detach was requested (explicit detach command only, not Ctrl-C)
-        $SmakCli::cli_owner = -1;  # Mark CLI as unowned
-        if ($own_server) {
-            print "Detaching from CLI (job server still running)...\n";
-        } else {
-            print "Detaching from job server...\n";
-        }
+	$ret = "stop";
     }
 
-    return $exit_requested ? 'stop' : 'detach';
+    return $ret;
 }
+
+our $busy = 0; # more to do, not just waiting.
+our $rp_pending = 0;
 
 sub reprompt()
 {
     # Send SIGWINCH to wake up readline and trigger redraw
-    # Only if we're not busy with a build
-    return if $busy;
-
-    my $pid = $SmakCli::cli_owner;
-    if ($pid >= 0) {
-        kill 'WINCH', $pid;
+    if ($busy) {
+	$rp_pending++;
+    } else {
+	my $pid = $SmakCli::cli_owner;
+	kill 'WINCH', $pid if ($pid >= 0);
     }
 }
 
@@ -2996,6 +3003,8 @@ sub enable_cli {
 
 sub cmd_build {
     my ($words, $socket, $opts, $state) = @_;
+    
+    $busy = 1;
 
     my @targets = @$words;
     if (@targets == 0) {
@@ -3009,9 +3018,6 @@ sub cmd_build {
             return;
         }
     }
-
-    enable_cli(0);
-    $busy = 1;
 
     if ($socket) {
         # Job server mode - submit jobs and wait for results
@@ -3074,7 +3080,9 @@ sub cmd_build {
     }
 
     $busy = 0;
-    enable_cli(1);
+    if ($rp_pending) {
+	reprompt();
+    }
 }
 
 sub cmd_rebuild {
@@ -5258,7 +5266,7 @@ sub run_job_master {
             vprint "Dispatched task $task_id to worker\n";
 
 	    if (! $silent_mode) {
-		print "$job->{command}\n";
+		print $stomp_prompt,"$job->{command}\n";
 	    }
 	    
             broadcast_observers("DISPATCHED $task_id $job->{target}");
@@ -6115,7 +6123,7 @@ sub run_job_master {
                 } elsif ($line =~ /^OUTPUT (.*)$/) {
                     my $output = $1;
                     # Always print to stdout (job master's stdout is inherited from parent)
-                    print "$output\n";
+                    print $stomp_prompt,"$output\n";
                     # Also forward to attached clients if any
                     print $master_socket "OUTPUT $output\n" if $master_socket;
